requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aioprocessing==2.0.1
aiosignal==1.3.1
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backoff==2.2.1
bitsandbytes==0.43.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decord==0.6.0
deepspeed==0.14.4
distro==1.9.0
docker-pycreds==0.4.0
draccus==0.8.0
einops==0.8.0
emoji==2.12.1
exceptiongroup==1.2.2
executing==2.0.1
facexlib==0.3.0
filelock==3.15.4
filterpy==1.4.5
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
grpcio==1.66.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
icecream==2.1.3
idna==3.7
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.4.0
janus==1.0.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mergedeep==1.3.4
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.41.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyyaml-include==1.4.1
qwen-vl-utils==0.0.4
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.11.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@6b7d64ac1c5d19ab4b9e99142a08e31741106e4e
transformers-stream-generator==0.0.5
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid_utils==0.9.0
wandb==0.17.7
wcwidth==0.2.13
weave==0.50.14
Werkzeug==3.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.20.0
